requirements.txt
streamlit
pymupdf
nltk
sentence-transformers
faiss-cpu
transformers
torch
